% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_txn_table.R
\name{create_txn_table}
\alias{create_txn_table}
\title{pivot table - transaction amounts by year and month}
\usage{
create_txn_table(x)
}
\description{
pivot table - transaction amounts by year and month
}
