% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_client_transactions.R
\name{load_client_transactions}
\alias{load_client_transactions}
\title{Load client data merged with transactions}
\usage{
load_client_transactions()
}
\description{
Load client data merged with transactions
}
